BJS2-40681 validation handlers in one class and renamed ProjectDfs to ProjectSearcher
