--- src/main/java/faang/school/projectservice/exception/handler/DataValidationHandler.java
@@ -1,18 +1,0 @@
-package faang.school.projectservice.exception.handler;
-
-import faang.school.projectservice.exception.DataValidationException;
-import org.springframework.core.Ordered;
-import org.springframework.core.annotation.Order;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-@RestControllerAdvice
-@Order(Ordered.HIGHEST_PRECEDENCE)
-public class DataValidationHandler {
-    @ExceptionHandler(DataValidationException.class)
-    public ResponseEntity<String> handleDataValidation(DataValidationException ex) {
-        return new ResponseEntity<>(ex.getMessage(), HttpStatus.BAD_REQUEST);
-    }
-}
