--- src/main/java/faang/school/projectservice/exception/handler/EntityNotFoundHandler.java
@@ -1,18 +1,0 @@
-package faang.school.projectservice.exception.handler;
-
-import jakarta.persistence.EntityNotFoundException;
-import org.springframework.core.Ordered;
-import org.springframework.core.annotation.Order;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-@RestControllerAdvice
-@Order(Ordered.HIGHEST_PRECEDENCE)
-public class EntityNotFoundHandler {
-    @ExceptionHandler(EntityNotFoundException.class)
-    public ResponseEntity<String> handleEntityNotFoundException(EntityNotFoundException ex) {
-        return new ResponseEntity<>(ex.getMessage(), HttpStatus.NOT_FOUND);
-    }
-}
